requirements.txt
requests
pandas
python-telegram-bot
ta
schedule